chore(be): stop sending access_token to the client
This was just my mistake when I first wrote this. I was new and didn't really
understand much of this. (Which is why I was writing it in the first place).

I am keeping expiresOn in there for now because the isInSession middleware
uses it to check if the token is expired and get a new one. This needs to be
rethought but I don't have much time for this atm.