feat: move login page to use react with index.ejs
Currently, we have two different .ejs templates. One just for the login page
and one for the rest of the app. Now we have react router, and things are more
put together let's remove the login.ejs template.

This way we don't have to maintain a separate css, and js for it (or combine it
with the react one). Less complex build process as well.